Add load_vocoder function to svs.py and use them throughout the code
